# Optim: MORI-IO eliminate repeated session construction via cache